Rename constant file
